+++ app/src/main/java/com/example/diningroomfactory/Models/ProductPair.kt
@@ -7,6 +7,6 @@ data class ProductPair (
     @PropertyName("product")
     var productId: String = "",
     @PropertyName("number")
-    var number: Int = 0
+    var number: Double = 0.0
 
 )
